code/requirements.txt
transformers==4.57.1
opencv-python==4.12.0.88
Pillow==11.3.0
numpy
accelarate==1.11.0